Rename configdialog to settingsdialog
