+++ GuiApp/GuiApp.pro
@@ -9,11 +9,11 @@ SOURCES    = main.cpp
 
 win32:RC_ICONS += icons/depthmapX.ico
 
-win32:Release:LIBS += -L../depthmapX/release -L../genlib/release -L../salalib/release -L../configdialog/release
-win32:Debug:LIBS += -L../depthmapX/debug -L../genlib/debug -L../salalib/debug -L../configdialog/debug
-!win32:LIBS += -L../depthmapX -L../genlib -L../salalib -L../configdialog
+win32:Release:LIBS += -L../depthmapX/release -L../genlib/release -L../salalib/release -L../settingsdialog/release
+win32:Debug:LIBS += -L../depthmapX/debug -L../genlib/debug -L../salalib/debug -L../settingsdialog/debug
+!win32:LIBS += -L../depthmapX -L../genlib -L../salalib -L../settingsdialog
 
-LIBS += -ldepthmapX -lsalalib -lgenlib -lconfigdialog
+LIBS += -ldepthmapX -lsalalib -lgenlib -lsettingsdialog
 
 !win32:!macx:LIBS += -L/usr/lib/i386-linux-gnu/
