+++ GuiApp/main.cpp
@@ -36,7 +36,7 @@
 int main(int argc, char *argv[])
 {
     Q_INIT_RESOURCE(resource);
-    Q_INIT_RESOURCE(configdialog);
+    Q_INIT_RESOURCE(settingsdialog);
 
     CoreApplication app(argc, argv);
