+++ settingsdialog/generalpage.cpp
