+++ settingsdialog/generalpage.h
