+++ settingsdialog/interfacepage.cpp
