+++ settingsdialog/settingsdialog.qrc
